backend/requirements.txt
fastapi
uvicorn[standard]
gunicorn

pydantic[email]

PyJWT
passlib[bcrypt]

httpx
beautifulsoup4

google-genai
stripe

sqlalchemy
psycopg2-binary
asyncpg

prometheus-fastapi-instrumentator

python-dotenv
websockets

# optional only if actually used
redis